requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
coverage==7.13.1
Django==4.2.27
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
iniconfig==2.3.0
joblib==1.5.3
librt==0.7.7
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nltk==3.9.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
pytokens==0.3.0
regex==2026.1.15
ruff==0.14.10
sqlparse==0.5.5
tqdm==4.67.1
types-python-dateutil==2.9.0.20251115
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
